+++ CHANGES.md
@@ -6,6 +6,11 @@ Version 0.9.5
 
 To be released.
 
+ -  Fixed a bug that did not properly receive block hashes after the chain has reorged.
+    [[#880], [#905]]
+
+[#905]: https://github.com/planetarium/libplanet/pull/905
+
 
 Version 0.9.4
 --------------